+++ RSCDLCEnabler/D3D/D3D.hpp
@@ -46,7 +46,7 @@ inline LPDIRECT3DTEXTURE9 Red, Green, Blue, Yellow;
 inline LPDIRECT3DTEXTURE9 pCurrTexture;
 inline LPDIRECT3DTEXTURE9 gradientTextureNormal, gradientTextureSeven, nonexistentTexture, additiveNoteTexture, ourTexture, twitchUserDefinedTexture;
 inline std::vector<LPDIRECT3DTEXTURE9> randomTextures(randomTextureCount);
-inline std::vector<LPDIRECT3DTEXTURE9> rainbowTextures(360.0f / rainbowSpeed);
+inline std::vector<LPDIRECT3DTEXTURE9> rainbowTextures((const unsigned int)(360.0f / rainbowSpeed));
 inline std::vector<ColorList> randomTextureColors;
 
 inline int selectedIdx = 0, counter = 0, currentRandomTexture = 0;
